pyproject.toml
[project]
name = "project-finforensic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.14",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.3",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.11",
    "google-api-python-client>=2.165.0",
    "google-genai>=1.7.0",
    "google-search-results>=2.4.2",
    "jinja2>=3.1.6",
    "langchain>=0.3.21",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.46",
    "langchain-deepseek>=0.1.2",
    "langchain-google-genai>=2.1.0",
    "langchain-groq>=0.3.0",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "lxml>=5.3.1",
    "markdown>=3.7",
    "markdown-it-py>=3.0.0",
    "markdownify>=1.1.0",
    "mistralai>=1.5.2",
    "mypy>=1.15.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "pydantic[email,timezone]>=2.10.6",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
    "youtube-transcript-api>=1.0.2",
]
